requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.9.0,<3
python-multipart==0.0.9
jinja2==3.1.4
Pillow==10.4.0
openai>=1.40.0,<2